requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
flask
requests
tqdm
scikit-learn
seaborn
matplotlib
pytest
httpx